Moved definition separate, also updated device list
