+++ src/app/devices/devices.component.html
@@ -1,15 +1,50 @@
 <div class="device-box-container" *ngFor="let device of devices">
   
-  <div class="device-box flex-container" >
-    <div class="info-box">
-      <i class="{{device.icon}}"></i>
-      <h2>{{device.GetTitle}}</h2>
-      <span class="description">Last connectivity: 1 minute ago</span>
+  <div class="device-box" >
+    <div class='flex-container'>
+      <div class="info-box">
+        <i class="{{device.icon}}"></i>
+        <h2>{{device.GetTitle}}</h2>
+        <span class="description">Last connectivity: 1 minute ago</span>
+      </div>
+      
+      <div class="io-overview">
+        <span class='text-info inputs'>{{device.Inputs.length}}<sup>Inputs</sup></span>
+        <span class="text-info outputs">{{device.Outputs.length}}<sup>Outputs</sup></span>
+      </div>
     </div>
+    <hr>
     
-    <div class="io-overview">
-      <span class='text-info inputs'>{{device.Inputs.length}}<sup>Inputs</sup></span>
-      <span class="text-info outputs">{{device.Outputs.length}}<sup>Outputs</sup></span>
-    </div>  
+    <!-- Input managing -->
+    <div class='flex-container vertical-container content-padding' *ngIf='device.Inputs.length'>
+        <div class="vertical-row">
+          <h3>Inputs</h3>
+        </div>
+        <div class="vertical-row">
+          <span class="button-container" *ngFor="let input of device.Inputs">
+            <label><input type="checkbox" class="ios-switch smallswitch" checked /><div><div></div></div></label>
+          </span>
+        </div>
+    </div>
+
+    <hr>
+    <!-- Outputs managing -->
+    <div class='flex-container vertical-container content-padding' *ngIf='device.Outputs.length'>
+        <div class="vertical-row">
+          <h3>Outputs</h3>
+        </div>
+        <div class="vertical-row">
+          <span class="button-container" *ngFor="let input of device.Outputs">
+            <label><input type="checkbox" class="ios-switch smallswitch" checked /><div><div></div></div></label>
+          </span>
+        </div>
+    </div>
+
   </div>
 </div>
+
+<br>
+<br>
+<pre style='font-size:10px;'>
+{{devices | json}}
+</pre>
