+++ src/app/shared/Definitions.ts
@@ -1,0 +1,21 @@
+/**
+ * A pin is any kind of I/O port of a device.
+ * it has a type, which can be input, or output
+ * also there is a unique id on device.
+ * value, is the current status of device, can be TRUE,FALSE for digital
+ * or can be a float number for analog data, such as temperature, humidity, light
+ */
+export interface IPin {
+  type: any,
+  id: any,
+  value: any
+}
+
+/**
+ * A device is consisting an array of pins.
+ */
+export interface IDevice {
+  uniqueid: string,
+  pins: Array<IPin>
+}
+
